# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CompetitionStatus = typeof CompetitionStatus[keyof typeof CompetitionStatus];


export const CompetitionStatus = {
  active: 'active',
  ended: 'ended',
  coming_soon: 'coming_soon',
} as const;

export interface Competition {
  id: number;
  title: string;
  description: string;
  prizeValue: number;
  ticketPrice: number;
  maxTickets: number;
  ticketsSold: number;
  status: CompetitionStatus;
  category: string;
  imageUrl: string;
  drawDate: string;
  isFeatured?: boolean;
  hasInstantWin?: boolean;
  createdAt: string;
}

export interface TicketPurchaseInput {
  /** @minimum 1 */
  quantity: number;
  email: string;
  name: string;
  selectedNumbers?: number[];
}

export interface TicketPurchase {
  id: number;
  competitionId: number;
  quantity: number;
  totalPrice: number;
  ticketNumbers: number[];
  email: string;
  name: string;
  createdAt: string;
}

export interface DrawResult {
  id: number;
  competitionId: number;
  competitionTitle: string;
  winnerName: string;
  winnerTicket: number;
  prize: string;
  prizeValue?: number;
  /** @nullable */
  imageUrl?: string | null;
  drawnAt: string;
}

export interface Winner {
  id: number;
  name: string;
  prize: string;
  amount: number;
  date: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  testimonial?: string | null;
}

export interface PlatformStats {
  totalWinners: number;
  totalPrizesAwarded: number;
  activeCompetitions: number;
  lowestTicketPrice: number;
  trustpilotRating: number;
  totalReviews: number;
}

export type ListCompetitionsParams = {
status?: ListCompetitionsStatus;
category?: string;
limit?: number;
offset?: number;
};

export type ListCompetitionsStatus = typeof ListCompetitionsStatus[keyof typeof ListCompetitionsStatus];


export const ListCompetitionsStatus = {
  active: 'active',
  ended: 'ended',
  coming_soon: 'coming_soon',
} as const;

export type ListDrawResultsParams = {
limit?: number;
};

export type ListWinnersParams = {
limit?: number;
};