# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Prize Competitions API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: competitions
    description: Competition operations
  - name: tickets
    description: Ticket purchase operations
  - name: draws
    description: Draw results operations
  - name: winners
    description: Winners operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /competitions:
    get:
      operationId: listCompetitions
      tags: [competitions]
      summary: List all competitions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, ended, coming_soon]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of competitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competition"

  /competitions/featured:
    get:
      operationId: getFeaturedCompetitions
      tags: [competitions]
      summary: Get featured/highlighted competitions
      responses:
        "200":
          description: Featured competitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competition"

  /competitions/stats:
    get:
      operationId: getCompetitionStats
      tags: [competitions]
      summary: Get platform statistics
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

  /competitions/{id}:
    get:
      operationId: getCompetition
      tags: [competitions]
      summary: Get a single competition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
        "404":
          description: Not found

  /competitions/{id}/tickets:
    post:
      operationId: purchaseTickets
      tags: [tickets]
      summary: Purchase tickets for a competition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketPurchaseInput"
      responses:
        "201":
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketPurchase"
        "400":
          description: Bad request

  /draws:
    get:
      operationId: listDrawResults
      tags: [draws]
      summary: List draw results
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Draw results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrawResult"

  /winners:
    get:
      operationId: listWinners
      tags: [winners]
      summary: List recent winners
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent winners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Winner"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Competition:
      type: object
      required:
        - id
        - title
        - description
        - prizeValue
        - ticketPrice
        - maxTickets
        - ticketsSold
        - status
        - category
        - imageUrl
        - drawDate
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        prizeValue:
          type: number
        ticketPrice:
          type: number
        maxTickets:
          type: integer
        ticketsSold:
          type: integer
        status:
          type: string
          enum: [active, ended, coming_soon]
        category:
          type: string
        imageUrl:
          type: string
        drawDate:
          type: string
        isFeatured:
          type: boolean
        hasInstantWin:
          type: boolean
        createdAt:
          type: string

    TicketPurchaseInput:
      type: object
      required:
        - quantity
        - email
        - name
      properties:
        quantity:
          type: integer
          minimum: 1
        email:
          type: string
        name:
          type: string
        selectedNumbers:
          type: array
          items:
            type: integer

    TicketPurchase:
      type: object
      required:
        - id
        - competitionId
        - quantity
        - totalPrice
        - ticketNumbers
        - email
        - name
        - createdAt
      properties:
        id:
          type: integer
        competitionId:
          type: integer
        quantity:
          type: integer
        totalPrice:
          type: number
        ticketNumbers:
          type: array
          items:
            type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string

    DrawResult:
      type: object
      required:
        - id
        - competitionId
        - competitionTitle
        - winnerName
        - winnerTicket
        - prize
        - drawnAt
      properties:
        id:
          type: integer
        competitionId:
          type: integer
        competitionTitle:
          type: string
        winnerName:
          type: string
        winnerTicket:
          type: integer
        prize:
          type: string
        prizeValue:
          type: number
        imageUrl:
          type: ["string", "null"]
        drawnAt:
          type: string

    Winner:
      type: object
      required:
        - id
        - name
        - prize
        - amount
        - date
      properties:
        id:
          type: integer
        name:
          type: string
        prize:
          type: string
        amount:
          type: number
        date:
          type: string
        imageUrl:
          type: ["string", "null"]
        testimonial:
          type: ["string", "null"]

    PlatformStats:
      type: object
      required:
        - totalWinners
        - totalPrizesAwarded
        - activeCompetitions
        - lowestTicketPrice
        - trustpilotRating
        - totalReviews
      properties:
        totalWinners:
          type: integer
        totalPrizesAwarded:
          type: number
        activeCompetitions:
          type: integer
        lowestTicketPrice:
          type: number
        trustpilotRating:
          type: number
        totalReviews:
          type: integer