# lib/api-zod/src/generated/types/competition.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitionStatus } from './competitionStatus';

export interface Competition {
  id: number;
  title: string;
  description: string;
  prizeValue: number;
  ticketPrice: number;
  maxTickets: number;
  ticketsSold: number;
  status: CompetitionStatus;
  category: string;
  imageUrl: string;
  drawDate: string;
  isFeatured?: boolean;
  hasInstantWin?: boolean;
  createdAt: string;
}