# lib/api-zod/src/generated/types/drawResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */

export interface DrawResult {
  id: number;
  competitionId: number;
  competitionTitle: string;
  winnerName: string;
  winnerTicket: number;
  prize: string;
  prizeValue?: number;
  /** @nullable */
  imageUrl?: string | null;
  drawnAt: string;
}