# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalWinners: number;
  totalPrizesAwarded: number;
  activeCompetitions: number;
  lowestTicketPrice: number;
  trustpilotRating: number;
  totalReviews: number;
}