# lib/api-zod/src/generated/types/ticketPurchase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketPurchase {
  id: number;
  competitionId: number;
  quantity: number;
  totalPrice: number;
  ticketNumbers: number[];
  email: string;
  name: string;
  createdAt: string;
}