# lib/api-zod/src/generated/types/ticketPurchaseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketPurchaseInput {
  /** @minimum 1 */
  quantity: number;
  email: string;
  name: string;
  selectedNumbers?: number[];
}