# lib/api-zod/src/generated/types/winner.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prize Competitions API
 * OpenAPI spec version: 0.1.0
 */

export interface Winner {
  id: number;
  name: string;
  prize: string;
  amount: number;
  date: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  testimonial?: string | null;
}